Initial commit
Added a working version of the game to GitHub to track changes. Added more blocks to place and an example house in the screenshots. Restructured the project and added comments on what each piece does. Added some files with TODO in the first line to guide me with next steps. Created a README.md file with simple instructions on how to play the game. The game has been tested on a MacBook Pro with a M2 processor running macOS and on a desktop PC with a Ryzen 7 2700X and a RX Vega 56 running Windows 11.